Add Function C Controller

+++ ppc1/ppc1/src/main/java/src/main/java/Solver.java
@@ -6,14 +6,14 @@ package src.main.java;
 public class Solver {
     private int Cap_Labor = 0;
     private int Cap_Grape = 0;
-    private double Prc_Rose = 0 ;
-    private double Prc_Noir = 0 ;
+    private float Prc_Rose = 0 ;
+    private float Prc_Noir = 0 ;
 
     private int Opt_Rose = 0;
     private int Opt_Noir = 0;
     //private int Production_Capacity_Limit = 0;
     //private double Man_Power_Limit = 0 ;
-    public Solver(int Cap_Labor,int Cap_Grape,double Prc_Rose,double Prc_Noir){
+    public Solver(int Cap_Labor,int Cap_Grape,float Prc_Rose,float Prc_Noir){
         this.Cap_Labor = Cap_Labor;
         this.Cap_Grape = Cap_Grape;
         this.Prc_Rose =  Prc_Rose;
@@ -23,7 +23,7 @@ public class Solver {
         //Man_Power_Limit = Num_week * 37.5;
     }
     public int[] Opt_Solution() {
-        double maxRevenue = 0;
+        float maxRevenue = 0;
         int opt_Rose = 0 ;
         int opt_Noir = 0 ;
         int i = 0 ;
@@ -39,7 +39,7 @@ public class Solver {
                     j= 0;
                     break;
                 }
-                double revenue = Calculation(i,j);
+                float revenue = Calculation(i,j);
                 if(maxRevenue<revenue) {
                     maxRevenue = revenue;
                     opt_Rose = i;
@@ -53,7 +53,7 @@ public class Solver {
         int[] result = {opt_Rose,opt_Noir};
         return result;
     }
-    public  double Calculation(int Num_Rose,int Num_Noir){
+    public  float Calculation(int Num_Rose,int Num_Noir){
         return Prc_Rose*Num_Rose + Prc_Noir*Num_Noir;
     }
 }
